Updated QueueService, added sending batch messages to rabbitmq, added queue with tokio::mpsc::channel, we are sending messages now with batch_send by 10 messages.

+++ src/domain/transaction.rs
@@ -32,8 +32,8 @@ pub struct Status {
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct Transaction {
     txid: String,
-    version: u8,
-    locktime: u16,
+    version: u64,
+    locktime: u64,
     vin: Vec<VectorInputs>,
     vout: Vec<VectorOutputs>,
     size: u32,
